+++ tests/test_func.py
@@ -1,0 +1,5 @@
+from qvarnmr.func import item
+
+
+def test_func_repr():
+    assert repr(item('id', 'value')) == "item('id', 'value')"